requirements.txt
requests
urllib3==1.26.6
websockets
python-telegram-bot